@@ -23,52 +23,33 @@
     ]
   },
   "dependencies": {
-    "autoprefixer": "^6.3.6",
-    "babel-core": "^6.7.7",
-    "babel-loader": "^6.2.4",
-    "body-parser": "~1.13.2",
     "classnames": "^2.2.3",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
-    "es6-promise": "^3.1.2",
-    "eslint-plugin-react": "^3.16.1",
-    "exports-loader": "^0.6.3",
     "express": "^4.13.4",
     "forever": "^0.15.1",
-    "html-webpack-plugin": "^1.7.0",
     "immutable": "^3.8.1",
-    "jquery": "^2.2.3",
-    "less": "^2.6.0",
-    "less-loader": "^2.2.3",
+    "js-cookie": "^2.1.3",
     "lodash": "^4.12.0",
-    "material-ui": "^0.14.4",
     "moment": "^2.13.0",
-    "morgan": "~1.6.1",
-    "node-jsx": "^0.13.3",
-    "node-libs-browser": "^0.5.3",
     "q": "^1.4.1",
     "ramda": "^0.17.1",
-    "react": "0.14.8",
-    "react-art": "^0.14.0",
-    "react-copy-to-clipboard": "^3.0.4",
-    "react-dom": "0.14.8",
-    "react-geosuggest": "^1.22.0",
+    "react": "^15.4.1",
+    "react-copy-to-clipboard": "^4.2.3",
+    "react-dom": "^15.4.1",
+    "react-geosuggest": "^1.25.1",
     "react-ink": "^5.0.2",
     "react-loader": "^2.4.0",
     "react-motion": "^0.4.3",
     "react-redux": "^4.4.5",
-    "react-router": "^2.3.0",
+    "react-router": "^2.8.1",
+    "react-router-redux": "^4.0.7",
     "react-scroll": "^1.0.5",
-    "react-tap-event-plugin": "^0.2.1",
     "redux": "^3.5.2",
     "redux-logger": "^2.6.1",
     "redux-thunk": "^2.1.0",
     "rx": "^3.1.2",
     "serve-favicon": "~2.3.0",
     "soundmanager2": "^2.97.20150601",
-    "style-loader": "^0.12.4",
-    "underscore": "^1.8.3",
-    "webpack": "^1.13.0"
+    "underscore": "^1.8.3"
   },
   "devDependencies": {
     "autoprefixer": "^6.3.3",
@@ -78,18 +59,19 @@
     "babel-plugin-transform-object-rest-spread": "^6.6.5",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "css-loader": "^0.26.1",
     "es6-promise": "^3.0.2",
-    "eslint-plugin-babel": "^3.1.0",
-    "eslint-plugin-react": "^3.2.3",
     "exports-loader": "^0.6.2",
     "file-loader": "^0.8.5",
-    "html-webpack-plugin": "^1.6.1",
     "imports-loader": "^0.6.5",
     "json-loader": "^0.5.3",
     "jsx-loader": "^0.13.2",
+    "less": "^2.7.1",
     "less-loader": "^2.2.2",
     "postcss-loader": "^0.8.2",
+    "react-addons-shallow-compare": "^15.4.1",
     "react-hot-loader": "^1.3.0",
+    "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
     "style-loader": "^0.12.3",
