@@ -11,64 +11,64 @@ const setsTooltip = 'Log in to see sets from artists you\'ve favorited';
 const eventsTooltip = 'Log in to see recommended events';
 
 export default class Home extends Base {
-	constructor(props) {
-		super(props);
-		// this.state = {
-		// 	disabled: true
-		// };
-	}
-	componentWillMount() {
-		const { push, loginStatus } = this.context;
-		push({ currentPage: 'Setmine' });
-		// if(loginStatus) {
-		// 	this.setState({ disabled: false });
-		// }
-	}
-	componentWillReceiveProps(nextProps, nextContext) {
-		// console.log(this.context.loginStatus, nextContext.loginStatus);
-		// if(nextContext.loginStatus) {
-		// 	this.setState({ disabled: false });
-		// }
-	}
-	componentDidMount() {
-		mixpanel.track("User Home Page Open");
-	}
-	render() {
-		return (
-			<div className='view'>
-				<Nav>
-					<IndexLink to='/home'>
-						<p>STREAM</p>
-						<Ink />
-					</IndexLink>
-					{this.context.loginStatus ? 
-						<Link to='/home/favorites'>
-							<p>FAVORITES</p>
-							<Ink />
-						</Link> : <DisabledTab tooltip={favoritesTooltip}>FAVORITES</DisabledTab>}
-					{this.context.loginStatus ? 
-						<Link to='/home/sets'>
-							<p>SETS</p>
-							<Ink />
-						</Link> : <DisabledTab tooltip={setsTooltip}>SETS</DisabledTab>}
-					{this.context.loginStatus ? 
-						<Link to='/home/events'>
-							<p>EVENTS</p>
-							<Ink />
-						</Link> : <DisabledTab tooltip={eventsTooltip}>EVENTS</DisabledTab>}
-				</Nav>
-				{
-					React.cloneElement(this.props.children, {
-						appState: this.props.appState
-					})
-				}
-			</div>
-		);
-	}
+    constructor(props) {
+        super(props);
+        // this.state = {
+        //  disabled: true
+        // };
+    }
+    componentWillMount() {
+        const { push, loginStatus } = this.context;
+        push({ currentPage: 'Setmine' });
+        // if(loginStatus) {
+        //  this.setState({ disabled: false });
+        // }
+    }
+    componentWillReceiveProps(nextProps, nextContext) {
+        // console.log(this.context.loginStatus, nextContext.loginStatus);
+        // if(nextContext.loginStatus) {
+        //  this.setState({ disabled: false });
+        // }
+    }
+    componentDidMount() {
+        mixpanel.track("User Home Page Open");
+    }
+    render() {
+        return (
+            <div className='view'>
+                <Nav>
+                    <IndexLink to='/home'>
+                        <p>STREAM</p>
+                        <Ink />
+                    </IndexLink>
+                    {this.context.loginStatus ? 
+                        <Link to='/home/favorites'>
+                            <p>FAVORITES</p>
+                            <Ink />
+                        </Link> : <DisabledTab tooltip={favoritesTooltip}>FAVORITES</DisabledTab>}
+                    {this.context.loginStatus ? 
+                        <Link to='/home/sets'>
+                            <p>SETS</p>
+                            <Ink />
+                        </Link> : <DisabledTab tooltip={setsTooltip}>SETS</DisabledTab>}
+                    {this.context.loginStatus ? 
+                        <Link to='/home/events'>
+                            <p>EVENTS</p>
+                            <Ink />
+                        </Link> : <DisabledTab tooltip={eventsTooltip}>EVENTS</DisabledTab>}
+                </Nav>
+                {
+                    React.cloneElement(this.props.children, {
+                        appState: this.props.appState
+                    })
+                }
+            </div>
+        );
+    }
 }
 
 Home.contextTypes =  {
-	push: React.PropTypes.func,
-	user: React.PropTypes.object,
-	loginStatus: React.PropTypes.bool
+    push: React.PropTypes.func,
+    user: React.PropTypes.object,
+    loginStatus: React.PropTypes.bool
 };
