@@ -1,25 +1,25 @@
-import React, {PropTypes} from 'react';
-import Loader from 'react-loader';
-import R from 'ramda';
-import api from '../services/api';
-import {DEFAULT_IMAGE} from '../constants/constants';
+import React, { PropTypes } from 'react'
+import Loader from 'react-loader'
+import api from '../services/api'
+import { DEFAULT_IMAGE } from '../constants/constants'
 
-import Base from './Base';
-import DetailHeader from './DetailHeader';
-// import SocialMediaLinks from './SocialMediaLinks';
-import ShuffleButton from './ShuffleButton';
-import Tabs from './Tabs';
-import Tab from './Tab';
+import Base from './Base'
+import DetailHeader from './DetailHeader'
+import ShuffleButton from './ShuffleButton'
+import Tabs from './Tabs'
+import Tab from './Tab'
+
+import { changeCurrentPage } from '../actions/environment'
 
 const tabStyle = {
 	position: 'relative',
 	top: 0
-};
+}
 
 export default class ArtistDetail extends Base {
 	constructor(props) {
-		super(props);
-		this.autoBind('getArtist');
+		super(props)
+		this.autoBind('getArtist')
 		this.state = {
 			loaded: false,
 			sets: [],
@@ -27,19 +27,18 @@ export default class ArtistDetail extends Base {
 			artistImage: DEFAULT_IMAGE,
 			setCount: 0,
 			eventCount: 0
-		};
+		}
 	}
 	componentWillMount() {
-		this.getArtist();
-		this.context.push({ currentPage: 'Artists' });
+		this.getArtist()
+		this.props.dispatch(changeCurrentPage('Artists'))
 	}
 	getArtist() {
-		let artist = this.props.params.artist;
-		let query = artist.split('_').join('%20');
+		const artist = this.props.params.artist
+		const query = artist.split('_').join('%20')
 
 		api.get(`artists/search/${query}`).then(payload => {
-			let a = payload.artists_search;
-			// this.context.push({ currentPage: a.artist });
+			const a = payload.artists_search
 
 			this.setState({
 				artist: a.artist,
@@ -48,17 +47,18 @@ export default class ArtistDetail extends Base {
 				artistImage: a.icon_image.imageURL,
 				setCount: a.set_count,
 				eventCount: a.event_count
-			});
+			})
 		}).then(() => {
-			this.setState({ loaded: true });
-		});
+			this.setState({ loaded: true })
+		})
 	}
 	render() {
-		const setText = this.state.setCount != 1 ? 'sets' : 'set';
-		const eventText = this.state.eventCount != 1 ? 'events' : 'event';
-		const artistInfo = `${this.state.setCount} ${setText} | ${this.state.eventCount} ${eventText}`;
-
-		const setIds = R.pluck('id', this.state.sets);
+		const setText = this.state.setCount != 1 ? 'sets' : 'set'
+		const eventText = this.state.eventCount != 1 ? 'events' : 'event'
+		const artistInfo = `${this.state.setCount} ${setText} | ${this.state.eventCount} ${eventText}`
+		const setIds = this.state.sets.map(set => {
+			return set.id
+		})
 
 		return (
 			<Loader loaded={this.state.loaded}>
@@ -66,7 +66,7 @@ export default class ArtistDetail extends Base {
 					<DetailHeader image={this.state.artistImage}>
 						<h3>{this.state.artist}</h3>
 						<h5>{artistInfo}</h5>
-						<ShuffleButton setIds={setIds} />
+						<ShuffleButton setIds={setIds} dispatch={this.props.dispatch} />
 					</DetailHeader>
 					<Tabs type='detail' style={tabStyle}>
 						<Tab to={`/artist/${this.props.params.artist}`} index>SETS</Tab>
@@ -80,10 +80,6 @@ export default class ArtistDetail extends Base {
 					}
 				</div>
 			</Loader>
-		);
+		)
 	}
 }
-
-ArtistDetail.contextTypes = {
-	push: PropTypes.func
-};
