@@ -1,16 +1,14 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react'
 
-const DisabledTab = ({children, tooltip}) => {
-	return (
-		<div className='disabled-tab' title={tooltip}>
-			<p>{children}</p>
-		</div>
-	);
+export default function DisabledTab({ children, tooltip }) {
+    return (
+        <div className='disabled-tab' title={tooltip}>
+            <p>{children}</p>
+        </div>
+    )
 }
 
 DisabledTab.propTypes = {
-	children: PropTypes.string.isRequired,
-	tooltip: PropTypes.string.isRequired
-};
-
-export default DisabledTab;
+    children: PropTypes.string.isRequired,
+    tooltip: PropTypes.string.isRequired
+}
