@@ -1,51 +1,61 @@
-import React, {PropTypes} from 'react';
-import Base from './Base';
-import CopyToClipboard from 'react-copy-to-clipboard';
-import { favoriteSet } from '../services/favoriteSet';
-import { shareToFacebook, shareToTwitter } from '../services/share';
+import React, { PropTypes } from 'react'
+import Base from './Base'
+import CopyToClipboard from 'react-copy-to-clipboard'
+import Icon from './Icon'
+
+import { favoriteSet } from '../services/favoriteSet'
+import { shareToFacebook, shareToTwitter } from '../services/share'
 
 export default class SetShare extends Base {
 	constructor(props) {
-		super(props);
-		this.autoBind('handleCopy', 'favoriteSet', 'shareToFacebook', 'shareToTwitter');
+		super(props)
+		this.autoBind('handleCopy', 'favoriteSet', 'shareToFacebook', 'shareToTwitter')
+		this.state = {
+			favorited: props.favorited
+		}
 	}
 	handleCopy() {
-		this.context.push({
-			snackbar: {
-				open: true,
-				message: 'Copied to Clipboard!'
-			}
-		});
+		// this.context.push({
+		// 	snackbar: {
+		// 		open: true,
+		// 		message: 'Copied to Clipboard!'
+		// 	}
+		// })
 	}
 	favoriteSet() {
-		var {user, loginStatus, push} = this.context;
-		if(!loginStatus) {
-			push({ showLogin: true });
-			return;
-		}
+		// const { user, loginStatus, push } = this.context
+		// if(!loginStatus) {
+		// 	push({ showLogin: true })
+		// 	return
+		// }
 
-		favoriteSet(this.props.id, user.id, push);
+		// favoriteSet(this.props.id, user.id, push)
+		this.setState({ favorited: !this.state.favorited }, () => {
+			console.log('favorited', this.props.id)
+			// dispatch(favoriteSet(id))
+		})
 	}
 	shareToFacebook() {
-		shareToFacebook(this.props.id, this.props.artistImage);
+		// shareToFacebook(this.props.id, this.props.artistImage)
 	}
 	shareToTwitter() {
-		shareToTwitter(this.props.id);
+		// shareToTwitter(this.props.id)
 	}
 	render() {
-		let favorite = this.props.favorited ? 'fa fa-fw fa-star' : 'fa fa-fw fa-star-o';
-		let playURL = `https://setmine.com/play/${this.props.id}`;
+		// const favorite = this.state.favorited ? 'favorite' : 'favorite_border'
+		const favorite = this.state.favorited ? 'heart' : 'heart-o'
+		const playURL = `https://setmine.com/play/${this.props.id}`
 
 		return (
-			<div id='SetShare' className='share flex-row'>
-				<i className={favorite} onClick={this.favoriteSet} title='Favorite Set' />
+			<div className='SetShare flex-row'>
+				<Icon onClick={this.favoriteSet} title='Favorite Set'>{favorite}</Icon>
 				<CopyToClipboard text={playURL} onCopy={this.handleCopy}>
-					<i className='fa fa-fw fa-clipboard' title='Copy to Clipboard' />
+					<Icon title='Copy to Clipboard'>clipboard</Icon>
 				</CopyToClipboard>
-				<i className='fa fa-fw fa-facebook' title='Share on Facebook' onClick={this.shareToFacebook} />
-				<i className='fa fa-fw fa-twitter' title='Share on Twitter' onClick={this.shareToTwitter} />
+				<Icon onClick={this.shareToFacebook} fixed title='Share on Facebook'>facebook</Icon>
+				<Icon onClick={this.shareToTwitter} fixed title='Share on Twitter'>twitter</Icon>
 			</div>
-		);
+		)
 	}
 }
 
@@ -53,10 +63,10 @@ SetShare.contextTypes = {
 	push: PropTypes.func,
 	user: PropTypes.object,
 	loginStatus: PropTypes.bool
-};
+}
 
 SetShare.propTypes = {
 	favorited: PropTypes.bool,
 	id: PropTypes.number,
 	artistImage: PropTypes.string
-};
+}
