@@ -1,32 +1,24 @@
-import React, {PropTypes} from 'react';
-import Base from './Base';
-import {playSet, updatePlayCount} from '../services/playerService';
+import React, { PropTypes } from 'react'
+import Base from './Base'
+import { shuffle } from '../actions/player'
 
 export default class ShuffleButton extends Base {
 	constructor(props) {
-		super(props);
-		this.autoBind('shuffle');
+		super(props)
+		this.autoBind('shuffle')
 	}
 	shuffle() {
-		const sets = this.props.setIds;
-		const random = Math.floor(Math.random() * (sets.length - 1));
-		const randomSetId = sets[random];
-		
-		playSet(randomSetId, this.context.push);
-		updatePlayCount(randomSetId, this.context.user.id);
+		const { sets, dispatch } = this.props
+		dispatch(shuffle(sets))
 	}
 	render() {
 		return (
-			<p id='DetailButton' onClick={this.shuffle}>SHUFFLE</p>
-		);
+			<p className='DetailButton' onClick={this.shuffle}>SHUFFLE</p>
+		)
 	}
 }
 
-ShuffleButton.contextTypes = {
-	push: PropTypes.func,
-	user: PropTypes.object
-};
-
 ShuffleButton.propTypes = {
-	setIds: PropTypes.array.isRequired
-};
+	setIds: PropTypes.array.isRequired,
+	dispatch: PropTypes.func.isRequired
+}
