@@ -1,7 +1,7 @@
-import React from 'react';
-import { createDevTools } from 'redux-devtools';
-import LogMonitor from 'redux-devtools-log-monitor';
-import DockMonitor from 'redux-devtools-dock-monitor';
+import React from 'react'
+import { createDevTools } from 'redux-devtools'
+import LogMonitor from 'redux-devtools-log-monitor'
+import DockMonitor from 'redux-devtools-dock-monitor'
 
 export default createDevTools(
     <DockMonitor 
@@ -10,4 +10,4 @@ export default createDevTools(
     >
         <LogMonitor />
     </DockMonitor>
-);
+)
