@@ -1,12 +1,22 @@
-import * as types from '../constants/actionTypes';
+import * as types from '../constants/actionTypes'
 
 const initialState = {
-    user: null
-};
+    user: null,
+    favorites: [],
+    sets: [],
+    events: []
+}
 
 export default function login(state = initialState, action) {
     switch(action.type) {
+        case types.LOGOUT:
+            return initialState
+        case types.RECEIVE_USER:
+            return {
+                ...state,
+                user: action.user
+            }
         default:
-            return state;
+            return state
     }
 }
