@@ -1,35 +1,43 @@
-import * as types from '../constants/actionTypes';
+import * as types from '../constants/actionTypes'
 
 const initialState = {
-    currentPage: 'Setmine Redux',
+    currentPage: 'Setmine',
     isMobile: false,
-    showNavbar: false,
-    loaded: true
-};
+    loaded: true,
+    playerVisible: false,
+    notification: '',
+    showModal: false,
+    showNotification: false
+}
 
 export default function environment(state = initialState, action) {
     switch (action.type) {
         case types.CHANGE_CURRENT_PAGE:
-            return Object.assign({}, state, {
+            return {
+                ...state,
                 currentPage: action.page
-            });
-
+            }
         case types.CHANGE_IS_MOBILE:
-            return Object.assign({}, state, {
+            return {
+                ...state,
                 isMobile: action.isMobile
-            });
-
-        case types.TOGGLE_NAVBAR:
-            return Object.assign({}, state, {
-                showNavbar: action.showNavbar
-            });
-
+            }
         case types.TOGGLE_LOADER:
-            return Object.assign({}, state, {
-                loaded: action.loaded
-            });
-            
+            return {
+                ...state,
+                loaded: !state.loaded
+            }
+        case types.TOGGLE_MODAL:
+            return {
+                ...state,
+                showModal: !state.showModal
+            }
+        case types.TOOGLE_PLAYER:
+            return {
+                ...state,
+                playerVisible: !state.playerVisible
+            }
         default:
-            return state;
+            return state
     }
 }
