@@ -9,15 +9,14 @@ const initialState = {
 export default function sets(state = initialState, action) {
     switch(action.type) {
         case types.RECEIVE_SETS:
-            return Object.assign({}, state, {
+            return {
+                ...state,
                 page: action.page,
                 sets: action.sets,
                 loaded: action.loaded
-            })
-
+            }
         case types.RESET_SETS:
             return initialState;
-            
         default:
             return state;
     }
